requirements.txt
# API clients
anthropic>=0.40.0
openai>=1.50.0

# Document processing
python-docx>=1.1.0
openpyxl>=3.1.0

# Environment
python-dotenv>=1.0.0

# Data
pandas>=2.2.0
